refactor: move `factories.py` to test folder
