--- questionpy_common/dev/__init__.py
@@ -1,3 +1,0 @@
-#  This file is part of QuestionPy. (https://questionpy.org)
-#  QuestionPy is free software released under terms of the MIT license. See LICENSE.md.
-#  (c) Technische Universität Berlin, innoCampus <info@isis.tu-berlin.de>
